+++ scglobals/static/tx/src/app/services/initialqueries.js
@@ -69,10 +69,10 @@ export default class InitialQueries {
   saveDefaultQueries(){
     let querytypes = [
       // {queryName: 'shipments-by-date', doc_type: 'shipment', queryEmit: 'doc.date,doc.total_value'},
-      {queryName: 'shipments-by-date', doc_type: 'shipment', queryEmit: '[doc.date,doc.from_location_name,doc.to_location_name],doc.total_value'},
+      {queryName: 'shipments-by-date', doc_type: 'shipment', queryEmit: 'emit([doc.from_location_name,doc.date,doc.from_location_name,doc.to_location_name],doc.total_value); emit([doc.to_location_name,doc.date,doc.from_location_name,doc.to_location_name],doc.total_value);'},
       // for sum of total value (when working, use sum & skip "count" of shipments)
-      {queryName: 'shipments-by-value', doc_type: 'shipment', queryEmit: '[doc.date,doc.from_location_name,doc.to_location_name], doc.total_value', reduce: '_sum'},
-      {queryName: 'shipments-by-count', doc_type: 'shipment', queryEmit: '[doc.date,doc.from_location_name,doc.to_location_name], doc.total_transactions', reduce: '_sum'}];
+      {queryName: 'shipments-by-value', doc_type: 'shipment', queryEmit: 'emit([doc.from_location_name,doc.date,doc.from_location_name,doc.to_location_name], doc.total_value); emit([doc.to_location_name,doc.date,doc.from_location_name,doc.to_location_name], doc.total_value);', reduce: '_sum'},
+      {queryName: 'shipments-by-count', doc_type: 'shipment', queryEmit: 'emit([doc.from_location_name,doc.date,doc.from_location_name,doc.to_location_name], doc.total_transactions); emit([doc.to_location_name,doc.date,doc.from_location_name,doc.to_location_name], doc.total_transactions);', reduce: '_sum'}];
     var addQueries = [];
     var addQuery;
     var functionString;
@@ -82,7 +82,7 @@ export default class InitialQueries {
         views: {}
       };
       addQuery.views[query.queryName] = {};
-      functionString = 'function(doc){if (doc.doc_type && doc.doc_type === "strReplaceQueryName"){emit(strReplaceQueryEmit);}}';
+      functionString = 'function(doc){if (doc.doc_type && doc.doc_type === "strReplaceQueryName"){strReplaceQueryEmit}}';
       functionString = functionString.replace(/strReplaceQueryEmit/g,query.queryEmit);
       addQuery.views[query.queryName]["map"] = functionString.replace(/strReplaceQueryName/g, query.doc_type);
       addQuery.views[query.queryName]["reduce"] = query.reduce || "_count";